Added day 8
More work on understanding dictionaries and calling elements of a dictionary, getting lists with an index.  Actual algorithm was pretty simple when based on the right data structure.

For part 2, assumed each ghost was on a long loop and then took the LCM of all the routes.  I didn't check the assumption they were valid loops - but turns out the answer was right - the 10e12 steps would have taken for ever to run as a group.