Day 9 Added
This meant playing with lists of lists, and using [-1] notation to find the last item in a list.

insert(0,x) was used to prepend to a list.

Hard part was discovering lists are generally passed by reference - so part 1 was altering our original analysis list we created, so repeating the work again for part 2 gave a different answer.

list.copy() solves for this.  As would have resetting the list and reading in a second time.