Day 11 Added First version used a real grid for all the galaxiues and expanded it then did the search to fidn the co-ordinates.  worked fine for an age 1 galaxy, but not when adding millions of NULL rows.
Instead, found all the galaxy co-ordinates first, then worked out which rows/columns were empty; and then added to the x/y of the galaxies for each row/column below it that was empty.

Finally solved Day 18 using a shoelace formula - something that calculates size based on the points of a polygon, not by drawing the polygon and filling it in.

Never worked out quite why the perimeter has a fudge factor of +1 to make the size work, but at this point - time to move on.

Learn more about grid maths, not so much about Python.  I think this could have been done via a class or two, but didn't really need it for the final calculation.